Pulled RoutingTableHandler interface out
